+++ index.html
@@ -146,8 +146,11 @@
     </section>
 
     
+    <!-- modules -->
+    <script src="js/modal.js" type="module"></script>
+    <script src="js/visit.js" type="module"></script>
     <!-- main js -->
-    <script src="js/app.js"></script>
+    <script src="js/app.js" type="module"></script>
     <!-- bootstrap -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous">
